Update to filters and models
Firstly, filters and models are now namespaced. This makes event
triggering far easier, because it uses the namespace. It also looks
cleaner.

Secondly, both routes and models can now register a filter. Both methods
are RESTful, which is the groundwork for the API service. A model can
specify an array of filters using the following methods :

- getFilters()
- postFilters()
- deleteFilters()
- putFilterS()

All filters are first registered in the FilterStack, which then
initiates them when they're all registered.

Further WebService changes include :

- Models are now automatically registered in the ObjectContainer
- Events have been removed (nothing can hook into them anyway)