Event triggers are variadic & use references
Events now use the name PHP 5.6 variadic feature. The first argument is
the name of the event, following arguments are then sent to the event
hook. Event hooks can also request for references, allowing you to edit
them within the scope of the object which called the event (particularly
useful in model events).